feat: modified error handler to return possible rules for a validation failed rules
